% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_raw_pbp.R
\name{save_raw_pbp}
\alias{save_raw_pbp}
\title{Query PBP FROM NFL API AND SAVE LOCALLY}
\usage{
save_raw_pbp(game_nflapi_id, game_nflverse_id, filepath)
}
\arguments{
\item{game_nflapi_id}{The game id of the nfl api}

\item{game_nflverse_id}{The game id of the nflverse}

\item{filepath}{A path to save raw pbp files to}
}
\value{
The filepath where the raw pbp was saved to
}
\description{
Query PBP FROM NFL API AND SAVE LOCALLY
}
